requirements.txt
playwright==1.50.0
Scrapy==2.12.0
scrapy-playwright==0.0.42
python-decouple==3.8
requests==2.32.3
